@@ -7,8 +7,7 @@ LAT code adapted from code provided by Simon Dicker.
 import logging
 from functools import lru_cache, partial
 import numpy as np
-from numpy.core.numeric import isscalar
-from scipy.interpolate import interp1d, bisplrep, bisplev
+from scipy.interpolate import make_interp_spline, bisplrep, bisplev
 from scipy.spatial.transform import Rotation as R
 from sotodlib import core
 from sotodlib.io.metadata import read_dataset
@@ -41,20 +40,12 @@ LAT_TUBES = {
     "o6": 11,
 }
 
+# fmt: off
 # SAT Optics
 # TODO: Maybe we want these to be provided in a config file?
-SAT_X = (0.0, 29.7580, 59.4574, 89.5745, 120.550, 152.821, 163.986, 181.218)
-SAT_LON = (
-    0.0,
-    0.0523597,
-    0.10471958,
-    0.15707946,
-    0.20943951,
-    0.26179764,
-    0.27925093,
-    0.30543087,
-)
-
+SAT_R_FP = (0.0, 29.7580, 59.4574, 89.5745, 120.550, 152.821, 163.986, 181.218)
+SAT_R_SKY = (0.0, 0.0523597, 0.10471958, 0.15707946, 0.20943951, 0.26179764, 0.27925093, 0.30543087)
+# fmt: on
 
 def _interp_func(x, y, spline):
     xr = np.atleast_1d(x).ravel()
@@ -128,8 +119,10 @@ def get_gamma(pol_xi, pol_eta):
     xi = pol_xi.reshape((-1, 2))
     eta = pol_eta.reshape((-1, 2))
 
-    d_xi = np.diff(xi, axis=1).ravel()
-    d_eta = np.diff(eta, axis=1).ravel()
+    q0 = quat.rotation_xieta(xi[:,0], eta[:,0])
+    q1 = quat.rotation_xieta(xi[:,1], eta[:,1])
+    dq = ~q0 * q1
+    d_xi, d_eta, _ = quat.decompose_xieta(dq)
 
     gamma = np.arctan2(d_xi, d_eta) % (2 * np.pi)
     return gamma
@@ -235,7 +228,7 @@ def ufm_to_fp(aman, x=None, y=None, pol=None, theta=0, dx=0, dy=0):
     return x_fp, y_fp, pol_fp
 
 
-def LAT_pix2sky(x, y, sec2xi, sec2eta, array2secx, array2secy, rot=0, opt2cryo=30.0):
+def LAT_pix2sky(x, y, sec2el, sec2xel, array2secx, array2secy, rot=0, opt2cryo=30.0):
     """
     Routine to map pixels from arrays to sky.
 
@@ -245,23 +238,23 @@ def LAT_pix2sky(x, y, sec2xi, sec2eta, array2secx, array2secy, rot=0, opt2cryo=3
 
         y: Y position on focal plane (currently zemax coord).
 
-        sec2xi: Function that maps positions on secondary to on sky xi. 
+        sec2el: Function that maps positions on secondary to on sky elevation. 
 
-        sex2eta: Function that maps positions on secondary to on sky eta.
+        sex2xel: Function that maps positions on secondary to on sky cross-elevation.
 
         array2secx: Function that maps positions on tube's focal plane to x position on secondary.
 
         array2secy: Function that maps positions on tube's focal plane to y position on secondary.
 
-        rot: Rotation about the line of site = xi - 60 - corotator.
+        rot: Rotation about the line of site = elev - 60 - corotator.
 
         opt2cryo: The rotation to get from cryostat coordinates to zemax coordinates (TBD, prob 30 deg).
 
     Returns:
 
-        xi: The on sky elevation in radians.
+        el: The on sky elevation in radians.
 
-        eta: The on sky eta in radians.
+        xel: The on sky cross-elevation radians.
     """
     d2r = np.pi / 180.0
     # TBD - put in check for MASK - values outside circle should not be allowed
@@ -272,15 +265,15 @@ def LAT_pix2sky(x, y, sec2xi, sec2eta, array2secx, array2secy, rot=0, opt2cryo=3
     xs = array2secx(xz, yz)
     ys = array2secy(xz, yz)
     # get into LAT zemax coord
-    # There is a 90 degree offset between how zemax coords and SO coords are defined
-    rot -= 90
+    # There is an offset between how zemax coords and SO coords are defined
+    rot += 180
     xrot = xs * np.cos(d2r * rot) - ys * np.sin(d2r * rot)
     yrot = ys * np.cos(d2r * rot) + xs * np.sin(d2r * rot)
     # note these are around the telescope boresight
-    xi = sec2xi(xrot, yrot)
-    eta = sec2eta(xrot, yrot)
+    el = sec2el(xrot, yrot)
+    xel = sec2xel(xrot, yrot)
 
-    return np.deg2rad(xi), np.deg2rad(eta)
+    return np.deg2rad(el), np.deg2rad(xel)
 
 
 @lru_cache(maxsize=None)
@@ -315,9 +308,9 @@ def LAT_optics(zemax_path):
 
     Returns:
 
-        sec2xi: Function that maps positions on secondary to on sky elevation
+        sec2el: Function that maps positions on secondary to on sky elevation
 
-        sex2eta: Function that maps positions on secondary to on sky eta.
+        sex2xel: Function that maps positions on secondary to on sky cross-elevation.
     """
     zemax_dat = load_zemax(zemax_path)
     try:
@@ -329,15 +322,15 @@ def LAT_optics(zemax_path):
     gi = np.where(LAT["mask"] != 0.0)
     x = LAT["x"][gi].ravel()
     y = LAT["y"][gi].ravel()
-    xi = LAT["elev"][gi].ravel()
-    eta = LAT["eta"][gi].ravel()
+    el = LAT["elev"][gi].ravel()
+    xel = LAT["xel"][gi].ravel()
 
-    s2e = bisplrep(x, y, xi, kx=3, ky=3)
-    sec2xi = partial(_interp_func, spline=s2e)
-    s2x = bisplrep(x, y, eta, kx=3, ky=3)
-    sec2eta = partial(_interp_func, spline=s2x)
+    s2e = bisplrep(x, y, el, kx=3, ky=3)
+    sec2el = partial(_interp_func, spline=s2e)
+    s2x = bisplrep(x, y, xel, kx=3, ky=3)
+    sec2xel = partial(_interp_func, spline=s2x)
 
-    return sec2xi, sec2eta
+    return sec2el, sec2xel
 
 
 @lru_cache(maxsize=None)
@@ -436,15 +429,17 @@ def LAT_focal_plane(aman, zemax_path, x=None, y=None, pol=None, roll=0, tube_slo
     if pol is None:
         pol = aman.focal_plane.pol_fp
 
-    sec2xi, sec2eta = LAT_optics(zemax_path)
+    sec2el, sec2xel = LAT_optics(zemax_path)
     array2secx, array2secy = LATR_optics(zemax_path, tube_slot)
 
-    xi, eta = LAT_pix2sky(x, y, sec2xi, sec2eta, array2secx, array2secy, roll)
+    el, xel = LAT_pix2sky(x, y, sec2el, sec2xel, array2secx, array2secy, roll)
+    xi, eta, _ = quat.decompose_xieta(quat.euler(1, np.deg2rad(90)) * quat.rotation_lonlat(-xel, el))
 
     pol_x, pol_y = gen_pol_endpoints(x, y, pol)
-    pol_xi, pol_eta = LAT_pix2sky(
-        pol_x, pol_y, sec2xi, sec2eta, array2secx, array2secy, roll
+    pol_el, pol_xel = LAT_pix2sky(
+        pol_x, pol_y, sec2el, sec2xel, array2secx, array2secy, roll
     )
+    pol_xi, pol_eta, _ = quat.decompose_xieta(quat.rotation_lonlat(-pol_xel, pol_el))
     gamma = get_gamma(pol_xi, pol_eta)
     if np.isscalar(xi):
         gamma = gamma[0]
@@ -463,21 +458,21 @@ def LAT_focal_plane(aman, zemax_path, x=None, y=None, pol=None, roll=0, tube_slo
 
 
 @lru_cache(maxsize=None)
-def sat_to_sky(x, theta):
+def sat_to_sky(r_fp, r_sky):
     """
-    Interpolate x and theta values to create mapping from SAT focal plane to sky.
+    Interpolate focal plane and on sky radius values to create mapping from SAT focal plane to sky.
     This function is a wrapper whose main purpose is the cache this mapping.
 
     Arguments:
-        x: X values in mm, should be all positive.
+        r_fp: Focal plane radius values in mm, should be all positive.
 
-        theta: Theta values in radians, should be all positive.
+        theta: On sky radius values in radians, should be all positive.
                Theta is defined by ISO coordinates.
 
     Return:
         sat_to_sky: Interp object with the mapping from the focal plane to sky.
     """
-    return interp1d(x, theta, fill_value="extrapolate")
+    return make_interp_spline(r_fp, r_sky)
 
 
 def SAT_focal_plane(aman, x=None, y=None, pol=None, roll=0, mapping_data=None):
@@ -497,7 +492,7 @@ def SAT_focal_plane(aman, x=None, y=None, pol=None, roll=0, mapping_data=None):
 
         roll: Rotation about the line of site = -1*boresight.
 
-        mapping_data: Tuple of (x, theta) that can be interpolated to map the focal plane to the sky.
+        mapping_data: Tuple of (r_fp, r_sky) that can be interpolated to map the focal plane to the sky.
                       Leave as None to use the default mapping.
 
     Returns:
@@ -519,32 +514,36 @@ def SAT_focal_plane(aman, x=None, y=None, pol=None, roll=0, mapping_data=None):
         pol = aman.focal_plane.pol_fp
 
     if mapping_data is None:
-        fp_to_sky = sat_to_sky(SAT_X, SAT_LON)
+        fp_to_sky = sat_to_sky(SAT_R_FP, SAT_R_SKY)
     else:
         mapping_data = (tuple(val) for val in mapping_data)
         fp_to_sky = sat_to_sky(*mapping_data)
 
-    # NOTE: lonlat coords are naturally centered at (1, 0, 0) and
-    #       xieta at (0, 0, 1). The euler angle below does this recentering
-    #       as well as flipping the sign of eta.
-    #       There is also a sign flip of xi that is supresses the factor of
-    #       -1 that would normally be applied when calculating lon since
-    #       it has the opposite sign as x.
-    #       The sign flips perform the flip about the origin from optics.
-    minus_lon = np.sign(x) * fp_to_sky(np.abs(x))
-    lat = np.sign(y) * fp_to_sky(np.abs(y))
-    _xi, _eta, _ = quat.decompose_xieta(
-        quat.euler(1, np.deg2rad(90)) * quat.rotation_lonlat(minus_lon, lat)
-    )
+    # Get things in polor coords 
+    r_fp = (x**2 + y**2)**.5
+    phi_fp = np.arctan2(y, x)
+
+    # Now to the sky
+    theta_iso = -fp_to_sky(r_fp)
+    phi_iso = np.pi/2 - phi_fp
+
+    # Flip about the origin for optics (180 deg rotation)
+    phi_iso += np.pi
+
+    # Now go to xieta
+    _xi, _eta, _ = quat.decompose_xieta(quat.rotation_iso(theta_iso, phi_iso))
+
+    # Apply roll
     xi = _xi * np.cos(np.deg2rad(roll)) - _eta * np.sin(np.deg2rad(roll))
     eta = _eta * np.cos(np.deg2rad(roll)) + _xi * np.sin(np.deg2rad(roll))
 
+    # Lets do gamma as a set of endpoints
     pol_x, pol_y = gen_pol_endpoints(x, y, pol)
-    pol_minus_lon = np.sign(pol_x) * fp_to_sky(np.abs(pol_x))
-    pol_lat = np.sign(pol_y) * fp_to_sky(np.abs(pol_y))
-    _xi, _eta, _ = quat.decompose_xieta(
-        quat.euler(1, np.deg2rad(90)) * quat.rotation_lonlat(pol_minus_lon, pol_lat)
-    )
+    pol_r = (pol_x**2 + pol_y**2)**.5
+    pol_phi = np.arctan2(pol_y, pol_x)
+    pol_theta_iso = -fp_to_sky(pol_r)
+    pol_phi_iso = np.pi/2 - pol_phi + np.pi
+    _xi, _eta, _ = quat.decompose_xieta(quat.rotation_iso(pol_theta_iso, pol_phi_iso))
     pol_xi = _xi * np.cos(np.deg2rad(roll)) - _eta * np.sin(np.deg2rad(roll))
     pol_eta = _eta * np.cos(np.deg2rad(roll)) + _xi * np.sin(np.deg2rad(roll))
     gamma = get_gamma(pol_xi, pol_eta)
@@ -574,6 +573,7 @@ def get_focal_plane(
     tube_slot=None,
     wafer_slot="ws0",
     config_path=None,
+    ufm_to_fp_pars=None,
     zemax_path=None,
     mapping_data=None,
     return_fp=False,
@@ -607,6 +607,9 @@ def get_focal_plane(
         wafer_slot: Which wafer slot to use.
 
         config_path: Path to the ufm_to_fp config file.
+
+        ufm_to_fp_pars: Loaded ufm_to_fp_parsm to focalplane params.
+                        If provided config_path is is ignored.
 
         zemax_path: Path to the data file from Zemax.
                     Only used by the LAT.
@@ -648,7 +651,8 @@ def get_focal_plane(
     if telescope_flavor not in ["LAT", "SAT"]:
         raise ValueError("Telescope should be LAT or SAT")
 
-    ufm_to_fp_pars = get_ufm_to_fp_pars(telescope_flavor, wafer_slot, config_path)
+    if ufm_to_fp_pars is None:
+        ufm_to_fp_pars = get_ufm_to_fp_pars(telescope_flavor, wafer_slot, config_path)
     x_fp, y_fp, pol_fp = ufm_to_fp(aman, x=x, y=y, pol=pol, **ufm_to_fp_pars)
     if telescope_flavor == "LAT":
         if zemax_path is None:
