@@ -65,7 +65,7 @@ def expand_RangesMatrix(flat_rm):
     if shape[0] == 0:
         return so3g.proj.RangesMatrix([], child_shape=shape[1:])
     # Otherwise non-trivial
-    count = np.product(shape[:-1])
+    count = np.prod(shape[:-1])
     start, stride = 0, count // shape[0]
     for i in range(0, len(ends), stride):
         _e = ends[i:i+stride] - start
